sliding & windows implementation unification
